+++ lib/http2/index.js
@@ -1,0 +1,2 @@
+export { Http2WebTransportClient } from './client.js'
+export { Http2WebTransportServer } from './server.js'
