+++ lib/server.js
@@ -1,12 +1,14 @@
 import { ReadableStream } from 'node:stream/web'
 import { HttpWTSession } from './session.js'
-import { wtrouter, quicheInit } from './native.js'
 import { WebTransportError } from './error.js'
-import { Http2WebTransportServer } from './http2/server.js'
-import { Http3WebTransportServerSocket } from './http3/serversocket.js'
+import { Http2WebTransportServer } from './http2/index.js'
 import { isIPv4 } from 'net'
 import { defer, logger } from './utils.js'
-import { HttpWebTransport } from './transport.js'
+import {
+  checkQuicheInit,
+  Http3WebTransportServer,
+  Http3WebTransportServerSocket
+} from './http3/index.js'
 
 const log = logger(`webtransport:httpserver(${process?.pid})`)
 
@@ -241,14 +243,7 @@ export class HttpServer {
         // @ts-ignore
         this.args?.reliability === 'both')
     ) {
-      if (!HttpWebTransport.quicheInited) {
-        quicheInit({
-          quicheLogVerbose: this.args?.quicheLogVerbose
-            ? this.args.quicheLogVerbose
-            : -1
-        })
-        HttpWebTransport.quicheInited = true
-      }
+      checkQuicheInit(args)
     }
 
     try {
@@ -258,7 +253,7 @@ export class HttpServer {
       switch (reliability) {
         case 'unreliableOnly':
           socket = new Http3WebTransportServerSocket(this.args)
-          this.transportInt = new wtrouter.Http3WebTransportServer(this.args)
+          this.transportInt = new Http3WebTransportServer(this.args)
           this.transportInt.stopServer = socket.stopServer.bind(socket)
           socket.init()
           socket.cobj = this.transportInt
@@ -274,7 +269,7 @@ export class HttpServer {
         case 'both':
           {
             socket = new Http3WebTransportServerSocket(this.args)
-            const server3 = new wtrouter.Http3WebTransportServer(this.args)
+            const server3 = new Http3WebTransportServer(this.args)
             server3.stopServer = socket.stopServer.bind(socket)
             socket.init()
             server3.socket = socket
