--- lib/transport.js
@@ -1,8 +1,0 @@
-/**
- * @typedef {import('./types').HttpServerInit} HttpServerInit
- * @typedef {import('./types').HttpClientInit} HttpClientInit
- */
-
-export class HttpWebTransport {
-  static quicheInited = false
-}
