+++ lib/webtransport.node.js
@@ -1,11 +1,13 @@
-import { Http3WebTransportClientSocket } from './http3/clientsocket.js'
+import {
+  checkQuicheInit,
+  Http3WebTransportClientSocket,
+  Http3WebTransportClient
+} from './http3/index.js'
 import { HttpWTSession } from './session.js'
 import { WebTransportBase } from './webtransportbase.js'
 import { HttpClient } from './client.js'
-import { quicheInit, wtrouter } from './native.js'
-import { Http2WebTransportClient } from './http2/client.js'
+import { Http2WebTransportClient } from './http2/index.js'
 import { logger } from './utils.js'
-import { HttpWebTransport } from './transport.js'
 
 const log = logger(`webtransportnode(${process?.pid})`)
 /**
@@ -66,19 +68,12 @@ export class WebTransport extends WebTransportBase {
       createUnreliableClient: function (client) {
         if (
           // @ts-ignore
-          !client.args?.forceReliable
+          !args?.forceReliable
         ) {
-          if (!HttpWebTransport.quicheInited) {
-            quicheInit({
-              quicheLogVerbose: args?.quicheLogVerbose
-                ? args.quicheLogVerbose
-                : -1
-            })
-            HttpWebTransport.quicheInited = true
-          }
+          checkQuicheInit(args)
         }
         const socket = new Http3WebTransportClientSocket(args)
-        const uclient = new wtrouter.Http3WebTransportClient(args)
+        const uclient = new Http3WebTransportClient(args)
         uclient.closeClientInt = uclient.closeClient
         uclient.closeClient = socket.closeClient.bind(socket)
         socket.init()
